# app/api/endpoints/assets.py
import csv
import io
import json
import logging
from datetime import datetime

from fastapi import APIRouter, Depends, File, Form, HTTPException, Query, Request, UploadFile, status
from fastapi.responses import (
    HTMLResponse,
    RedirectResponse,
    Response,
    StreamingResponse,
)
from pydantic import TypeAdapter, ValidationError
from sqlalchemy.exc import SQLAlchemyError
from sqlalchemy.ext.asyncio import AsyncSession

from app.api.deps import (
    get_current_superuser_from_session,
    get_current_user_from_session,
)
from app.db.database import get_db
from app.flash import flash, get_flashed_messages
from app.models.user import User
from app.schemas.asset import AssetCreate, AssetResponse, AssetUpdate
from app.schemas.component import ComponentItem
from app.services.component_service import ComponentService
from app.services.device_service import DeviceService
from app.services.exceptions import (
    DeletionError,
    DeviceNotFoundException,
    DuplicateDeviceError,
    NotFoundError,
)
from app.templating import templates
from app.utils.helpers import safe_int

logger = logging.getLogger(__name__)
router = APIRouter()


def get_device_service() -> DeviceService:
    return DeviceService()


@router.get('/assets-list', response_class=HTMLResponse, name='read_assets')
async def read_assets(
    request: Request,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    page: int = Query(1, ge=1),
    page_size: int = Query(20, ge=1, le=100),
    search: str | None = Query(None),
    asset_type_id: str | None = Query(None),
    status_id: str | None = Query(None),
    department_id: str | None = Query(None),
    location_id: str | None = Query(None),
    manufacturer_id: str | None = Query(None),
    employee_id: str | None = Query(None),
    supplier_id: str | None = Query(None),
    tag_id: str | None = Query(None),
    sort_by: str | None = Query(None),
    sort_order: str = Query('asc'),
    current_user: User = Depends(get_current_user_from_session),
):
    """Отображает список активов с фильтрацией, сортировкой и пагинацией."""
    try:
        filters_dict = {
            'search': search,
            'asset_type_id': safe_int(asset_type_id),
            'status_id': safe_int(status_id),
            'department_id': safe_int(department_id),
            'location_id': safe_int(location_id),
            'manufacturer_id': safe_int(manufacturer_id),
            'employee_id': safe_int(employee_id),
            'supplier_id': safe_int(supplier_id),
            'tag_id': safe_int(tag_id),
        }

        (
            paginated_devices_db,
            total_devices,
        ) = await device_service.get_devices_with_filters(
            db=db,
            page=page,
            page_size=page_size,
            sort_by=sort_by,
            sort_order=sort_order,
            **filters_dict,
        )
        paginated_devices = [
            AssetResponse.model_validate(d, from_attributes=True)
            for d in paginated_devices_db
        ]
        form_data = await device_service.get_all_dictionaries_for_form(db)
        query_params = request.query_params._dict.copy()
        query_params.pop('page', None)

        context = {
            'request': request,
            'devices': paginated_devices,
            'total_devices': total_devices,
            'page': page,
            'page_size': page_size,
            'total_pages': (total_devices + page_size - 1) // page_size,
            'title': 'Список активов',
            **form_data,
            'endpoint_name': 'read_assets',
            'filters': {
                **filters_dict,
                'page_size': page_size,
                'sort_by': sort_by,
                'sort_order': sort_order,
            },
            'query_params': query_params,
        }
        return templates.TemplateResponse('assets_list.html', context)
    except SQLAlchemyError as e:
        logger.error(f'Ошибка при загрузке списка активов: {e}', exc_info=True)
        raise HTTPException(
            status_code=500, detail='Ошибка базы данных при загрузке активов.'
        )


@router.get("/dashboard", response_class=HTMLResponse, name="dashboard")
async def dashboard(
    request: Request,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user_from_session),
):
    """Отображает аналитический дашборд."""
    return templates.TemplateResponse(
        "dashboard.html",
        {"request": request, "title": "Дашборд - IT Asset Tracker"},
    )


@router.get('/export/csv', name='export_assets_csv')
async def export_assets_csv(
    request: Request,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    search: str | None = Query(None),
    asset_type_id: str | None = Query(None),
    status_id: str | None = Query(None),
    department_id: str | None = Query(None),
    location_id: str | None = Query(None),
    manufacturer_id: str | None = Query(None),
    current_user: User = Depends(get_current_user_from_session),
):
    """Экспортирует отфильтрованный список активов в CSV файл."""
    try:
        filters_dict = {
            'search': search,
            'asset_type_id': safe_int(asset_type_id),
            'status_id': safe_int(status_id),
            'department_id': safe_int(department_id),
            'location_id': safe_int(location_id),
            'manufacturer_id': safe_int(manufacturer_id),
        }
        devices_db, _ = await device_service.get_devices_with_filters(
            db=db, page=1, page_size=1_000_000, **filters_dict
        )
        devices = [
            AssetResponse.model_validate(d, from_attributes=True) for d in devices_db
        ]
        output = io.StringIO()
        output.write('\ufeff')
        writer = csv.writer(output)
        headers = [
            'ID',
            'Инвентарный номер',
            'Серийный номер',
            'MAC-адрес',
            'Тип',
            'Производитель',
            'Модель',
            'Статус',
            'Отдел',
            'Локация',
            'Сотрудник',
            'Дата покупки',
            'Окончание гарантии',
        ]
        writer.writerow(headers)
        for device in devices:
            writer.writerow(
                [
                    device.id,
                    device.inventory_number,
                    device.serial_number,
                    device.mac_address,
                    device.asset_type.name,
                    device.device_model.manufacturer.name,
                    device.device_model.name,
                    device.status.name,
                    device.department.name if device.department else '',
                    device.location.name if device.location else '',
                    f'{device.employee.last_name} {device.employee.first_name}'
                    if device.employee
                    else '',
                    device.purchase_date,
                    device.warranty_end_date,
                ]
            )
        output.seek(0)
        filename = f"assets_{datetime.now().strftime('%Y%m%d_%H%M%S')}.csv"
        return StreamingResponse(
            io.StringIO(output.getvalue()),
            media_type='text/csv',
            headers={'Content-Disposition': f'attachment; filename="{filename}"'},
        )
    except Exception as e:
        logger.error(f'Ошибка при экспорте в CSV: {e}', exc_info=True)
        flash(request, f'Ошибка при экспорте в CSV: {e}', 'danger')
        return RedirectResponse(
            url=request.url_for('read_assets'), status_code=status.HTTP_303_SEE_OTHER
        )


@router.get('/add', response_class=HTMLResponse, name='add_asset_form')
async def add_asset_form(
    request: Request,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_user_from_session),
):
    submitted_data = {}
    validation_errors = {}
    flashed_messages = get_flashed_messages(request)
    for msg in flashed_messages:
        if msg.get('category') == 'validation':
            if msg['message'].get('submitted_data'):
                submitted_data = msg['message']['submitted_data']
            if msg['message'].get('errors'):
                validation_errors = msg['message']['errors']

    asset = type('Asset', (), submitted_data)() if submitted_data else None
    try:
        form_data_for_selects = await device_service.get_all_dictionaries_for_form(db)
        all_tags = await device_service.get_all_tags(db)
    except Exception as e:
        logger.error(
            f"Ошибка при загрузке данных для формы 'add_asset': {e}", exc_info=True
        )
        raise HTTPException(
            status_code=500, detail='Ошибка при загрузке данных для формы.'
        )

    context = {
        'request': request,
        'title': 'Добавить актив',
        **form_data_for_selects,
        'all_tags': all_tags,
        'asset': asset,
        'errors': validation_errors,
        'messages': flashed_messages,
    }
    return templates.TemplateResponse('add_asset.html', context)


@router.post('/create', response_class=RedirectResponse, name='create_asset')
async def create_asset(
    request: Request,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_superuser_from_session),
):
    form_data = await request.form()
    form_dict = dict(form_data)
    tag_ids_str = form_data.getlist('tag_ids')
    tag_ids = [int(tag_id) for tag_id in tag_ids_str if tag_id.isdigit()]
    form_dict['tag_ids'] = tag_ids

    # Логируем полученные данные для отладки
    logger.debug('Получены данные формы создания актива:')
    logger.debug(f'  Все ключи: {list(form_data.keys())}')
    logger.debug(f'  form_dict: {form_dict}')
    logger.debug(
        f'  Обязательные поля: name={form_dict.get("name")}, '
        f'asset_type_id={form_dict.get("asset_type_id")}, '
        f'device_model_id={form_dict.get("device_model_id")}, '
        f'status_id={form_dict.get("status_id")}'
    )

    try:
        asset_data = AssetCreate(**form_dict)
        await device_service.create_device(db=db, asset_data=asset_data, user_id=current_user.id)
        flash(request, 'Актив успешно добавлен!', 'success')
        return RedirectResponse(
            url=request.url_for('read_assets'), status_code=status.HTTP_303_SEE_OTHER
        )
    except ValidationError as e:
        logger.warning(f'Ошибка валидации при создании актива: {e.errors()}')
        error_messages = []
        for error in e.errors():
            field = error['loc'][0] if error['loc'] else 'unknown'
            msg = error['msg']
            error_messages.append(f'{field}: {msg}')
        flash(
            request,
            {
                'errors': {'general': 'Пожалуйста, заполните все обязательные поля'},
                'submitted_data': form_dict,
            },
            'validation',
        )
    except (DuplicateDeviceError, NotFoundError) as e:
        logger.warning(f'Ошибка бизнес-логики при создании актива: {e}')
        flash(
            request,
            {'errors': {'general': str(e)}, 'submitted_data': form_dict},
            'validation',
        )
    except Exception as e:
        logger.error(f'Непредвиденная ошибка при создании актива: {e}', exc_info=True)
        flash(
            request,
            {
                'errors': {'general': f'Произошла непредвиденная ошибка: {e}'},
                'submitted_data': form_dict,
            },
            'validation',
        )
    return RedirectResponse(
        url=request.url_for('add_asset_form'), status_code=status.HTTP_303_SEE_OTHER
    )


@router.post('/import', name='import_asset_from_agent')
async def import_asset_from_agent(
    file: UploadFile,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_superuser_from_session),
):
    """
    Создает актив на основе JSON-отчета агента.
    """
    try:
        new_device = await device_service.create_from_agent_data(
            file=file, session=db, user_id=current_user.id
        )
        await db.commit()
        return {
            'status': 'success',
            'device_id': new_device.id,
            'redirect_url': f'/edit/{new_device.id}?tab=components',
        }
    except Exception as e:
        logger.error(f'Ошибка при импорте актива: {e}', exc_info=True)
        # Если это HTTPException, прокидываем его, иначе 500
        if isinstance(e, HTTPException):
            raise e
        raise HTTPException(status_code=500, detail=f'Ошибка импорта: {e!s}')


@router.get('/edit/{device_id}', response_class=HTMLResponse, name='edit_asset')
async def edit_asset(
    request: Request,
    device_id: int,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_user_from_session),
):
    device = await device_service.get_device_with_relations(db, device_id)
    if not device:
        raise HTTPException(status_code=404, detail='Устройство не найдено')
    try:
        form_data = await device_service.get_all_dictionaries_for_form(db)
    except Exception as e:
        logger.error(
            f'Ошибка при загрузке данных для формы редактирования: {e}', exc_info=True
        )
        raise HTTPException(
            status_code=500, detail='Ошибка при загрузке данных для формы.'
        )
    existing_tags_data = [{'id': tag.id, 'name': tag.name} for tag in device.tags]
    return templates.TemplateResponse(
        'edit_asset.html',
        {
            'request': request,
            'device': device,
            **form_data,
            'existing_tags_data': existing_tags_data,
            'title': f'Редактировать актив #{device.id}',
        },
    )


@router.post(
    '/update/{device_id}', response_class=RedirectResponse, name='update_asset'
)
async def update_asset(
    request: Request,
    device_id: int,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_superuser_from_session),
):
    form_data = await request.form()
    form_dict = dict(form_data)
    tag_ids_str = form_data.getlist('tag_ids')
    tag_ids = [int(tag_id) for tag_id in tag_ids_str if tag_id.isdigit()]
    form_dict['tag_ids'] = tag_ids
    try:
        update_data = AssetUpdate(**form_dict)
        updated_device = await device_service.update_device_with_audit(
            db=db, device_id=device_id, update_data=update_data, user_id=current_user.id
        )
        if not updated_device:
            raise DeviceNotFoundException('Актив не найден')
        flash(request, 'Актив успешно обновлен!', 'success')
        return RedirectResponse(
            url=request.url_for('read_assets'), status_code=status.HTTP_303_SEE_OTHER
        )
    except (DeviceNotFoundException, DuplicateDeviceError, NotFoundError) as e:
        logger.warning(f'Ошибка бизнес-логики при обновлении актива {device_id}: {e}')
        flash(
            request,
            {'errors': {'general': str(e)}, 'submitted_data': form_dict},
            'validation',
        )
    except Exception as e:
        logger.error(f'Ошибка при обновлении актива {device_id}: {e}', exc_info=True)
        flash(
            request,
            {
                'errors': {'general': f'Произошла непредвиденная ошибка: {e}'},
                'submitted_data': form_dict,
            },
            'validation',
        )
    return RedirectResponse(
        url=request.url_for('edit_asset', device_id=device_id),
        status_code=status.HTTP_303_SEE_OTHER,
    )


@router.post('/delete/{device_id}', name='delete_asset')
async def delete_asset(
    request: Request,
    device_id: int,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_superuser_from_session),
) -> Response:
    try:
        await device_service.delete_device_with_audit(
            db=db, device_id=device_id, user_id=current_user.id
        )
        flash(request, 'Актив успешно удален.', 'success')
    except DeviceNotFoundException as e:
        logger.warning(f'Попытка удаления несуществующего актива {device_id}: {e}')
        flash(request, 'Ошибка: Актив для удаления не найден.', 'danger')
    except DeletionError as e:
        logger.warning(f'Ошибка удаления актива {device_id} из-за зависимостей: {e}')
        flash(request, str(e), 'danger')
    except Exception as e:
        logger.error(f'Ошибка при удалении актива {device_id}: {e}', exc_info=True)
        flash(request, 'Произошла непредвиденная ошибка при удалении.', 'danger')
    return RedirectResponse(
        url=request.url_for('read_assets'), status_code=status.HTTP_303_SEE_OTHER
    )


@router.post('/assets/bulk-delete', name='bulk_delete_assets')
async def bulk_delete_assets(
    request: Request,
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_superuser_from_session),
) -> Response:
    form_data = await request.form()
    device_ids = [int(id) for id in form_data.getlist('device_ids')]
    referer = request.headers.get('referer') or request.url_for('read_assets')
    if not device_ids:
        flash(request, 'Не выбрано ни одного актива для удаления.', 'warning')
        return RedirectResponse(referer, status_code=status.HTTP_303_SEE_OTHER)
    try:
        deleted_count, errors = await device_service.bulk_delete_devices(
            db=db, device_ids=device_ids, user_id=current_user.id
        )
        if deleted_count > 0:
            flash(request, f'Успешно удалено {deleted_count} активов.', 'success')
        if errors:
            error_message = 'Не удалось удалить некоторые активы: ' + ', '.join(errors)
            flash(request, error_message, 'danger')
    except Exception as e:
        logger.error(f'Ошибка при массовом удалении активов: {e}', exc_info=True)
        flash(request, f'Произошла ошибка при удалении: {e}', 'danger')
    return RedirectResponse(referer, status_code=status.HTTP_303_SEE_OTHER)


@router.post('/assets/bulk-update', name='bulk_update_assets')
async def bulk_update_assets(
    request: Request,
    device_ids_json: str = Form(..., alias='device_ids_json'),
    db: AsyncSession = Depends(get_db),
    device_service: DeviceService = Depends(get_device_service),
    current_user: User = Depends(get_current_superuser_from_session),
    status_id: int | None = Form(None),
    department_id: int | None = Form(None),
    location_id: int | None = Form(None),
) -> Response:
    referer = request.headers.get('referer') or request.url_for('read_assets')
    try:
        device_ids = json.loads(device_ids_json)
        if not isinstance(device_ids, list) or not device_ids:
            flash(request, 'Не выбрано ни одного актива для обновления.', 'warning')
            return RedirectResponse(referer, status_code=status.HTTP_303_SEE_OTHER)
        update_data = {
            'status_id': status_id,
            'department_id': department_id,
            'location_id': location_id,
        }
        filtered_update_data = {k: v for k, v in update_data.items() if v is not None}
        if not filtered_update_data:
            flash(request, 'Не выбрано ни одного поля для обновления.', 'warning')
            return RedirectResponse(referer, status_code=status.HTTP_303_SEE_OTHER)
        updated_count = await device_service.bulk_update_devices(
            db=db, device_ids=device_ids, update_data=filtered_update_data, user_id=current_user.id
        )
        flash(request, f'Успешно обновлено {updated_count} активов.', 'success')
    except Exception as e:
        logger.error(f'Ошибка при массовом обновлении: {e}', exc_info=True)
        flash(request, f'Произошла ошибка при обновлении: {e}', 'danger')
    return RedirectResponse(referer, status_code=status.HTTP_303_SEE_OTHER)


def _normalize_component_data(data: dict | list) -> list[dict]:
    """Нормализует входящие данные в плоский список."""
    normalized_items = []
    if isinstance(data, dict):
        for key, value in data.items():
            if isinstance(value, list):
                for item in value:
                    if isinstance(item, dict):
                        if 'type' not in item:
                            if key in ('cpu', 'ram', 'storage', 'gpu', 'motherboard'):
                                item['type'] = key
                        normalized_items.append(item)
    elif isinstance(data, list):
        normalized_items = data
    return normalized_items


@router.post('/{asset_id}/components/upload', name='upload_components')
async def upload_components(
    asset_id: int,
    file: UploadFile = File(...),
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_superuser_from_session),
):
    """
    Загружает список компонентов из JSON файла.
    """
    if not file.filename.endswith('.json'):
        raise HTTPException(status_code=400, detail='Файл должен быть JSON')

    try:
        content = await file.read()
        data = json.loads(content)
    except json.JSONDecodeError:
        raise HTTPException(status_code=400, detail='Некорректный JSON')

    if not isinstance(data, (dict, list)):
        raise HTTPException(status_code=400, detail='JSON должен быть объектом или списком')

    normalized_items = _normalize_component_data(data)

    try:
        adapter = TypeAdapter(list[ComponentItem])
        items = adapter.validate_python(normalized_items)
    except ValidationError as e:
        errors = []
        for err in e.errors():
            loc = '->'.join(str(x) for x in err['loc'])
            errors.append(f"{loc}: {err['msg']}")
        raise HTTPException(status_code=422, detail=f'Ошибка валидации: {"; ".join(errors)}')

    await ComponentService.sync_components(db, asset_id, items)
    await db.commit()

    return {'status': 'ok', 'count': len(items)}